Uploaded draft tileset
The tileset was created using Tiled, which has a Unity plugin.
bar.tmx is the main project file, bar.tsx is the tileset, and bar.png is the image source for the tileset. They reference each other in that order (you may have to tidy up the references).
bar.tmx contains four layers. Three are cosmetic and one is a collision map. The collision map uses two collision tiles - one full tile, and one partial tile.